+++ cmd/titania.go/option.go
@@ -8,6 +8,13 @@ import (
 	"strings"
 )
 
+var (
+	quiet       = false
+	prettyprint = false
+	tmpdir      = ""
+	langs       = []string{}
+)
+
 func version() {
 	fmt.Printf("titania.go %s\n", VERSION)
 }
@@ -22,8 +29,10 @@ directories:
 options:
   -h, --help                   show this help message
   -v, --version                show version
+      --pretty                 pretty print output json
       --quiet                  quiet log
       --lang=lang1[,lang2,...] language[s] to test
+      --tmpdir=DIRNAME         set a directory where temporary files are put
 `)
 }
 
@@ -37,44 +46,55 @@ func cleanerpath(pwd, directory string) (string, error) {
 }
 
 // オプション解析
-func optparse() ([]string, []string, bool) {
+func optparse() []string {
 	var (
 		args     = make([]string, 0)
-		langs    = make([]string, 0)
 		dirnames = make([]string, 0)
-		quiet    = false
+		used     = -1
 	)
 
-	for _, arg := range os.Args[1:] {
-		if arg == "--help" || arg == "-h" {
+	pwd, _ := os.Getwd()
+
+	for i, arg := range os.Args[1:] {
+		if !strings.HasPrefix(arg, "-") {
+			if i != used {
+				args = append(args, arg)
+			}
+			continue
+		}
+
+		if strings.HasPrefix(arg, "--lang=") {
+			langs = strings.Split(strings.TrimPrefix(arg, "--lang="), ",")
+			continue
+		} else if strings.HasPrefix(arg, "--tmpdir=") {
+			tmpdir, _ = cleanerpath(pwd, strings.TrimPrefix(arg, "--tmpdir="))
+			continue
+		}
+
+		switch arg {
+		case "--help", "-h":
 			usage()
 			os.Exit(0)
-		} else if arg == "--version" || arg == "-v" {
+		case "--version", "-v":
 			version()
 			os.Exit(0)
-		} else if arg == "--quiet" {
+		case "--quiet":
 			quiet = true
-		} else if strings.HasPrefix(arg, "--lang=") {
-			langs = strings.Split(strings.Replace(arg, "--lang=", "", 1), ",")
-		} else if strings.HasPrefix(arg, "-") {
-			println("Unknown option:", arg)
+		case "--lang":
+			langs = strings.Split(strings.TrimPrefix(os.Args[i+2], "--lang="), ",")
+			used = i + 1
+		case "--tmpdir":
+			tmpdir, _ = cleanerpath(pwd, strings.TrimPrefix(os.Args[i+2], "--tmpdir="))
+			used = i + 1
+		default:
+			println("No such option:", arg)
 			usage()
 			os.Exit(1)
-		} else {
-			args = append(args, arg)
 		}
-	}
-
-	pwd, err := os.Getwd()
-	if err != nil {
-		panic(err)
 	}
 
 	if len(args) == 0 {
-		entries, err := ioutil.ReadDir(pwd)
-		if err != nil {
-			panic(err)
-		}
+		entries, _ := ioutil.ReadDir(pwd)
 
 		for _, entry := range entries {
 			if entry.IsDir() {
@@ -85,14 +105,11 @@ func optparse() ([]string, []string, bool) {
 	} else {
 
 		for _, directory := range args {
-			dirname, err := cleanerpath(pwd, directory)
-			if err != nil {
-				panic(err)
-			}
+			dirname, _ := cleanerpath(pwd, directory)
 			dirnames = append(dirnames, dirname)
 		}
 
 	}
 
-	return dirnames, langs, quiet
+	return dirnames
 }