+++ pkg/tester/test_target.go
@@ -1,17 +1,16 @@
 package tester
 
 import (
-	"io/ioutil"
 	"path/filepath"
 
-	"github.com/cwd-k2/titania.go/internal/pkg/langtype"
+	"github.com/cwd-k2/titania.go/pkg/runner"
 )
 
 type TestTarget struct {
-	Name       string
-	Language   string
-	SourceCode []byte
-	Expect     string
+	Name     string
+	Language string
+	FileName string
+	Expect   string
 }
 
 type TestTargetConfig struct {
@@ -19,10 +18,9 @@ type TestTargetConfig struct {
 	Expect  string `json:"expect"`
 }
 
-// Create []*TestTarget
 // This can return an empty slice.
 // All errors are logged but ignored.
-func MakeTestTargets(basepath string, configs []TestTargetConfig) []*TestTarget {
+func ReadTestTargets(basepath string, configs []TestTargetConfig) []*TestTarget {
 	targets := make([]*TestTarget, 0)
 
 	for _, config := range configs {
@@ -38,14 +36,8 @@ func MakeTestTargets(basepath string, configs []TestTargetConfig) []*TestTarget 
 		}
 
 		for _, filename := range filenames {
-			language := langtype.LangType(filename)
+			language := runner.LangType(filename)
 			if language == "plain" || len(languages) > 0 && !acceptable(languages, language) {
-				continue
-			}
-
-			sourceCodeBS, err := ioutil.ReadFile(filename)
-			if err != nil {
-				logger.Printf("%+v\n", err)
 				continue
 			}
 
@@ -54,7 +46,7 @@ func MakeTestTargets(basepath string, configs []TestTargetConfig) []*TestTarget 
 				name = filename
 			}
 
-			targets = append(targets, &TestTarget{name, language, sourceCodeBS, expect})
+			targets = append(targets, &TestTarget{name, language, filename, expect})
 		}
 	}