+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `simplejson` subpackage.
+- Added a functionality to store intermediate output files to a temporary directory.
+- Added `--temdir=` option to change the temporary directory.
+
+### Change
+
+- `paizaio` package was renamed to `runner`.
+- `pio` command was renamed to `piorun`.
+- Now output json is not pretty-printed by default. You can use `--pretty` flag to pretty-print.
+- `internal/pkg/pretty` package are now public.
+- `internal/pkg/langtype` were merged into `runner` package.
+
 ## [0.6.2] - 2021-02-17
 
 ### Added